# docs/api-spec.yaml
openapi: 3.1.0
info:
  title: Parcel Intelligence API
  version: 0.1.0
paths:
...
  /api/v1/grievances:
    post:
      tags:
      - grievances
      summary: Submit Grievance
      description: Submit a citizen land grievance or dispute report.
      operationId: submit_grievance_api_v1_grievances_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GrievanceCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GrievanceResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - grievances
      summary: List Grievances
      description: Retrieve citizen grievances with role-based visibility.
      operationId: list_grievances_api_v1_grievances_get
      security:
      - HTTPBearer: []
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GrievanceResponse'
                title: Response List Grievances Api V1 Grievances Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/grievances/{grievance_id}:
    patch:
      tags:
      - grievances
      summary: Update Grievance
      description: Update grievance resolution status and assign officer (revenue
        officers only).
      operationId: update_grievance_api_v1_grievances__grievance_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: grievance_id
        in: path
        required: true
        schema:
          type: integer
          title: Grievance Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GrievanceStatusUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GrievanceResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/dev/api-keys:
    post:
      tags:
      - developer
      summary: Generate Api Key
      description: Generate a new Developer Sandbox API Key for external integrations.
      operationId: generate_api_key_api_v1_dev_api_keys_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiKeyCreateRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyCreateResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/public/parcels/{ulpin}:
    get:
      tags:
      - developer
      summary: Get Public Parcel Data
      description: Public, sanitized parcel lookup endpoint (excludes citizen PII
        and sensitive legal encumbrances).
      operationId: get_public_parcel_data_api_v1_public_parcels__ulpin__get
      parameters:
      - name: ulpin
        in: path
        required: true
        schema:
          type: string
          title: Ulpin
      - name: x-api-key
        in: header
        required: true
        schema:
          type: string
          description: Developer Sandbox API Key
          title: X-Api-Key
        description: Developer Sandbox API Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicParcelResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ApiKeyCreateRequest:
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 3
          title: Name
          description: Developer or application name
        rate_limit_per_minute:
          anyOf:
          - type: integer
            maximum: 1000.0
            minimum: 1.0
          - type: 'null'
          title: Rate Limit Per Minute
          default: 60
      type: object
      required:
      - name
      title: ApiKeyCreateRequest
    ApiKeyCreateResponse:
      properties:
        id:
          type: integer
          title: Id
        name:
          type: string
          title: Name
        raw_api_key:
          type: string
          title: Raw Api Key
        rate_limit_per_minute:
          type: integer
          title: Rate Limit Per Minute
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - name
      - raw_api_key
      - rate_limit_per_minute
      - created_at
      title: ApiKeyCreateResponse
...
    GrievanceCreate:
      properties:
        ulpin:
          type: string
          title: Ulpin
        category:
          type: string
          maxLength: 100
          minLength: 3
          title: Category
        description:
          type: string
          minLength: 10
          title: Description
      type: object
      required:
      - ulpin
      - category
      - description
      title: GrievanceCreate
    GrievanceResponse:
      properties:
        id:
          type: integer
          title: Id
        parcel_id:
          type: integer
          title: Parcel Id
        ulpin:
          type: string
          title: Ulpin
        citizen_user_id:
          type: integer
          title: Citizen User Id
        citizen_name:
          type: string
          title: Citizen Name
        assigned_officer_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Assigned Officer Id
        category:
          type: string
          title: Category
        description:
          type: string
          title: Description
        status:
          type: string
          title: Status
        resolution_notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Resolution Notes
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - parcel_id
      - ulpin
      - citizen_user_id
      - citizen_name
      - category
      - description
      - status
      - created_at
      - updated_at
      title: GrievanceResponse
    GrievanceStatusUpdate:
      properties:
        status:
          type: string
          pattern: ^(submitted|under_review|hearing_scheduled|resolved|rejected)$
          title: Status
        resolution_notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Resolution Notes
        assigned_officer_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Assigned Officer Id
      type: object
      required:
      - status
      title: GrievanceStatusUpdate
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PublicParcelResponse:
      properties:
        ulpin:
          type: string
          title: Ulpin
        survey_number:
          anyOf:
          - type: string
          - type: 'null'
          title: Survey Number
        khasra_number:
          anyOf:
          - type: string
          - type: 'null'
          title: Khasra Number
        land_use_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Land Use Type
        area_sqm:
          anyOf:
          - type: number
          - type: 'null'
          title: Area Sqm
        state:
          anyOf:
          - type: string
          - type: 'null'
          title: State
        district:
          anyOf:
          - type: string
          - type: 'null'
          title: District
        village:
          anyOf:
          - type: string
          - type: 'null'
          title: Village
      type: object
      required:
      - ulpin
      title: PublicParcelResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer